requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.4
aiosignal==1.3.1 ; python_version >= '3.7'
algoliasearch==3.0.0
async-timeout==4.0.2 ; python_version >= '3.6'
asyncio==3.4.3
attrs==23.1.0 ; python_version >= '3.7'
backports.cached-property==1.0.2 ; python_full_version >= '3.6.0'
binapy==0.6.0 ; python_version >= '3.7' and python_version < '4.0'
blinker==1.6.2 ; python_version >= '3.7'
cachecontrol==0.13.0 ; python_version >= '3.6'
cachetools==5.3.1 ; python_version >= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
cryptography==41.0.1 ; python_version >= '3.7'
firebase-admin==6.1.0
flask==2.3.2
frozenlist==1.3.3 ; python_version >= '3.7'
furl==2.1.3
google-api-core[grpc]==2.11.0 ; platform_python_implementation != 'PyPy'
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.11.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1 ; platform_python_implementation != 'PyPy'
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2 ; python_version >= '3.7'
grpcio-status==1.54.2
gunicorn==20.1.0
httplib2==0.22.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.6.0 ; python_version < '3.10'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
jwskate==0.3.0 ; python_version >= '3.7' and python_version < '4.0'
loguru==0.7.0
markupsafe==2.1.3 ; python_version >= '3.7'
msgpack==1.0.5
multidict==6.0.4 ; python_version >= '3.7'
numpy==1.23.4
oauthlib==3.2.2 ; python_version >= '3.6'
orderedmultidict==1.0.1
packaging==23.1 ; python_version >= '3.7'
pandas==1.5.1
proto-plus==1.22.2 ; python_version >= '3.6'
protobuf==4.23.2 ; python_version >= '3.7'
pyasn1==0.5.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pyjwt[crypto]==2.7.0 ; python_version >= '3.7'
pyparsing==3.0.9 ; python_version >= '3.1'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3
requests==2.28.2
requests-oauth2client==1.1.0
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
setuptools==67.8.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==4.6.3 ; python_version >= '3.7'
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==1.26.16 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.3.4 ; python_version >= '3.8'
wheel==0.40.0
yarl==1.9.2 ; python_version >= '3.7'
zipp==3.15.0 ; python_version >= '3.7'
